challenge/requirements.txt
arrow==0.14.2
asn1crypto==0.24.0
cffi==1.12.3
Click==7.0
cryptography==2.7
Django==2.2.2
django-jsonfield==1.2.0
django-rest-knox==4.1.0
djangorestframework==3.9.4
Flask==1.0.3
geos==0.2.2
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.4
MarkupSafe==1.1.1
numpy==1.16.4
Pillow==6.0.0
postgis==1.0.4
proj==0.1.0
psycopg2==2.8.3
pycparser==2.19
pygdal==2.4.1.4
pyproj==2.2.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
Werkzeug==0.15.4
djangorestframework-gis
